Reorganized some files
