--- backend/src/main/java/com/IpTutor/Backend/authTest/SecurityConfig.java
@@ -1,38 +1,0 @@
-package com.IpTutor.Backend.authTest;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity(securedEnabled = true, jsr250Enabled = true)
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final AuthenticationProvider authenticationProvider;
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-      return http.authorizeHttpRequests(auth -> auth
-              .requestMatchers("/settings").authenticated()
-              .requestMatchers("/auth/test/settings").authenticated()
-              .anyRequest().permitAll()
-              )
-              .authenticationProvider(authenticationProvider)
-              .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-              .csrf(AbstractHttpConfigurer::disable)
-              .formLogin(form -> form.defaultSuccessUrl("/", true).loginPage("/login"))
-              .logout(config -> config.logoutSuccessUrl("/").logoutUrl("/logout"))
-              .build();
-    }
-
-}
