+++ backend/src/main/java/com/IpTutor/Backend/authentication/JwtService.java
@@ -1,4 +1,4 @@
-package com.IpTutor.Backend.authTest;
+package com.IpTutor.Backend.authentication;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
-import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
