+++ backend/src/main/java/com/IpTutor/Backend/config/SecurityConfig.java
@@ -1,5 +1,6 @@
-package com.IpTutor.Backend.authTest;
+package com.IpTutor.Backend.config;
 
+import com.IpTutor.Backend.authentication.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
