+++ backend/src/main/java/com/IpTutor/Backend/controller/AccountController.java
@@ -5,7 +5,6 @@ import com.IpTutor.Backend.dto.AccountRequestDTO;
 import com.IpTutor.Backend.dto.AccountDTO;
 import com.IpTutor.Backend.dto.LoginResponseDTO;
 import com.IpTutor.Backend.service.AccountService;
-import com.IpTutor.Backend.authTest.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
